Update and rename roller.ts to roller.js
